issue : session date attribute : int32 -> int256 date attribute + front get and set date

+++ contracts/Session.sol
@@ -7,19 +7,19 @@ import "hardhat/console.sol";
 contract Session 
 {
     uint16 private sessionId ; 
-    uint32 private date ; 
+    uint256 private date ; 
     uint16 private levelId ; 
     
     constructor(uint16 _sessionId)
     {
         sessionId = _sessionId ;
     }
-    function setDate(uint32 _date) public 
+    function setDate(uint256 _date) public 
     {
         date = _date;
     }
 
-    function getDate() view public returns(uint32)
+    function getDate() view public returns(uint256)
     {
         return date ; 
     }
